notification_service/requirements.txt
confluent-kafka==2.6.0
daphne==4.1.2
Django==5.1.3
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pillow==11.0.0
PyJWT==2.9.0
python-decouple==3.8
psycopg2==2.9.10
channels==4.2.0
channels_redis==4.2.1
pyfcm==2.0.7
firebase-admin==6.6.0
celery
django-celery-beat
django-webpush==0.3.6
elasticsearch==8.5.0